requirements.txt
asgiref==3.6.0
Django==4.1.6
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-taggit==3.1.0
Pillow==9.4.0
shortuuid==1.0.11
sqlparse==0.4.3
tzdata==2022.7
